backend/requirements.txt
flask==3.1.3
flask-sqlalchemy==3.1.1
langchain==1.2.15
langchain-openai==1.1.12
